requirements.txt
fastapi
uvicorn
python-dotenv
google-adk
google-genai
pydantic